PWN-6051 - Update Int to Float in user balance event

+++ app/src/main/java/org/p2p/wallet/home/analytics/HomeAnalytics.kt
@@ -3,6 +3,7 @@ package org.p2p.wallet.home.analytics
 import org.p2p.wallet.common.analytics.Analytics
 import org.p2p.wallet.common.analytics.constants.EventNames.HOME_USER_AGGREGATE_BALANCE
 import org.p2p.wallet.common.analytics.constants.EventNames.HOME_USER_HAS_POSITIVE_BALANCE
+import java.math.BigDecimal
 
 class HomeAnalytics(private val tracker: Analytics) {
 
@@ -11,12 +12,12 @@ class HomeAnalytics(private val tracker: Analytics) {
         logUserHasPositiveBalanceEvent(hasPositiveBalance)
     }
 
-    fun logUserAggregateBalanceProperty(usdBalance: Int) {
+    fun logUserAggregateBalanceProperty(usdBalance: BigDecimal) {
         tracker.setUserPropertyOnce(HOME_USER_AGGREGATE_BALANCE, usdBalance.toString())
         logUserAggregateBalanceEvent(usdBalance)
     }
 
-    private fun logUserAggregateBalanceEvent(usdBalance: Int) {
+    private fun logUserAggregateBalanceEvent(usdBalance: BigDecimal) {
         tracker.logEvent(
             HOME_USER_AGGREGATE_BALANCE,
             arrayOf("Usd_Balance" to usdBalance)
